requirements.txt
python-dotenv
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
langgraph
chromadb
sentence-transformers
pydantic
